Prime Generator PRIME1 SPOJ
This gives the solution to the PRIME1 problem on SPOJ. It can be done through C with usual prime finding technique( skipping even numbers and checking till n/2). However, Python being a slower language than C requires to implement a better algorithm to solve this problem. I have used Segmented Sieve method which is a version of Sieve of Eratosthenes.
Segmented Sieve algorithm provides all the prime numbers till a given number but we required prime numbers in a given interval. Hence, the algorithm of Segmented Sieve has been modified to meet the requirements.